formatting and modularising
