+++ .gitignore
@@ -3,3 +3,4 @@
 *.png
 *.pyc
 *.spydata
+*.html
