+++ poderballz.py
@@ -1,5 +1,12 @@
 # -*- coding: utf-8 -*-
 """
+Calculate Winston Luts results for multiple targets in a Dynamic HTT.
+
+Used alongside the a phantom with multiple ball bearing targets and an MLC 
+based plan with DCAT like apertures centred on the targets. The analysis will 
+provide details on the accuracy with which the planned apetures are delivered
+on treatment for a VMAT field with dynamic MLC and gantry rotation.
+
 .. module:: poderballz
    :platform: Windows
    :synopsis: grab some balls.
@@ -14,27 +21,50 @@ import matplotlib.pyplot as plt
 import pandas
 import glob
 import os
+import cv2
 import imageio
 
 from pathlib import Path
-from skimage import color
 from operator import itemgetter
 from skimage.transform import hough_circle, hough_circle_peaks
 from skimage.feature import canny
 from skimage.measure import regionprops, label
 from skimage.filters import threshold_otsu
+from poderprogressbar import update_progress
+from poderPlot import plot_against_gantry, boxplot, plot_coords_on_images
 
-import cv2
 
-import scipy.stats as stats
+def sparse_image(img,cropx,cropy):
+    """
+    Crop the imge and make it sparse.
+    
+    Parameters
+    ----------
+    img : numpy array
+        image to make sparse.
+    cropx : int
+        number pixels in the y direction.
+    cropy : int
+        numnber of pixels in the x direction.
 
+    Returns
+    -------
+    TYPE
+        DESCRIPTION.
+
+    """
+    y,x = img.shape
+    startx = x//2-(cropx//2)
+    starty = y//2-(cropy//2)    
+    return img[starty:starty+cropy,startx:startx+cropx]
 
 def get_apeture_centroids(image, binary, num_of_balls):
     """
-    Calculate the locations of all the mlc defined aperture positions in
-    the image. The location will be returned as the coordinates of the centre 
+    Calculate the locations of all the mlc defined aperture positions.
+    
+    The location will be returned as the coordinates of the centre 
     of mass of the aperture.
-
+    
     Parameters
     ----------
     image : numpy array
@@ -48,7 +78,6 @@ def get_apeture_centroids(image, binary, num_of_balls):
         list of apertures found in increasing y coordingates.
         
     """
-    
     label_image = label(binary)
     apertures = regionprops(label_image)
     apertures.sort(key=lambda a: a.convex_area, reverse=True)
@@ -61,7 +90,7 @@ def get_apeture_centroids(image, binary, num_of_balls):
     return centroids
 
 def get_ball_positions(image, binary, num_of_balls):
-    '''
+    """
     Return the positions of all the balls found in the image.
 
     Parameters
@@ -74,7 +103,7 @@ def get_ball_positions(image, binary, num_of_balls):
     ball_positions : TYPE
         DESCRIPTION.
 
-    '''
+    """
     edges = canny(image, sigma=3, low_threshold=5, high_threshold=10, mask=binary)
     hough_radii = np.arange(4, 15)
     hough_res = hough_circle(edges, hough_radii)
@@ -89,193 +118,9 @@ def get_ball_positions(image, binary, num_of_balls):
     #TODO - we will have to generalise this for geometry agnostic behaviour
     return ball_positions
 
-
-
-def sparse_image(img,cropx,cropy):
-    y,x = img.shape
-    startx = x//2-(cropx//2)
-    starty = y//2-(cropy//2)    
-    return img[starty:starty+cropy,startx:startx+cropx]
-
-
-#progress bar
-def update_progress(progress, subtext):
-    '''
-    Creates a progress bar in standard output.
-
-    Parameters
-    ----------
-    progress : int, float
-        DESCRIPTION.
-
-    Returns
-    -------
-    None.
-
-    '''
-    barLength = 20 # Modify this to change the length of the progress bar
-    status = ""
-    if isinstance(progress, int):
-        progress = float(progress)
-    if not isinstance(progress, float):
-        progress = 0
-        status = "error: progress var must be float\r\n"
-    if progress < 0:
-        progress = 0
-        status = "Halt...\r\n"
-    if progress >= 1:
-        progress = 1
-        status = "Done...\r\n"
-    block = int(round(barLength*progress))
-    text = "\rPercent: [{0}] {1}% {2} {3}".format( "#"*block + 
-                                                  "-"*(barLength-block),
-                                                  int(progress*100),
-                                                  status,
-                                                  subtext)
-    sys.stdout.write(text)
-    sys.stdout.flush()
-    
-
-# Plotting the balls
-def plot_against_gantry(what):
-    """
-    
-    Plots ball positions over gantry agnles
-    
-
-    Returns
-    -------
-    None.
-
-    """
-
-    
-    fig, (ax1, ax2) = plt.subplots(2, 1)
-    fig.suptitle('PoderBallz A tale of four balls {}'.format(what))
-    
-    colours = ['darkred', 'darkblue', 'darkgreen', 'darkslategrey']
-    # need more colours for more balls.
-    
-    # num_of_balls = 1 # comment out this line to print all balls
-    
-    for i in range(1,num_of_balls+1):
-        df.plot(kind="scatter", x='Gantry', y=(what, i, 'x'), s=1,
-                color=colours[i-1], label='ball {}'.format(i), ax=ax1)
-        
-        df.plot(kind="scatter", x='Gantry', y=(what, i, 'y'), s=1,
-                color=colours[i-1], label='ball {}'.format(i), ax=ax2)
-    
-    ax1.set_ylabel('X position (mm)')  
-    ax2.set_ylabel('Y position (mm)')
-    
-    ax1.legend(title="None", fontsize="xx-small", loc="upper right")
-    ax2.get_legend().remove()
-    
-    
-    plt.show()
-    
-    
-
-def plot_coords_on_images(what, range_of_images):
-    """
-    
-    Plots the coordinates of balls and apertures on the images.
-    
-    Parameters
-    ----------
-    image : numpy array
-        EPID image on which to show ball positions.
-    apertures : list of tuples
-        apeture positions.
-    balls : list of tuples 
-        ball positions.
-
-    Returns
-    -------
-    None.
-
-    """
-    if what == 'drrballs':
-        folder = drrfolder
-        aperture_dflabel = 'DRRApertures'
-        balls_dflabel = 'DRRBalls'
-    elif what == 'drrape':
-        folder = mlcfolder
-        aperture_dflabel = 'DRRApertures'
-        balls_dflabel = 'DRRBalls'
-    elif what == 'epid':
-        folder = epidfolder
-        aperture_dflabel = 'EPIDApertures'
-        balls_dflabel = 'EPIDBalls'
-        
-    
-    
-    for i in range_of_images:
-        # test plotting to see if the coordinates makes sense
-        filename = df.loc[i, 'filename'].values[0]
-        filename = folder / filename
-        im = imageio.imread(filename)
-        im = np.array(im)
-        
-        if what == 'epid':
-            balls = df.loc[i, balls_dflabel].values
-            apertures = df.loc[i, aperture_dflabel].values
-        elif what == 'drr' or 'drrape':
-            balls = df.loc[i, balls_dflabel].values
-            apertures = df.loc[i, aperture_dflabel].values
-
-        
-        #convert to format required by plotting function
-        it = iter(balls)
-        balls = [*zip(it,it)]
-        
-        it = iter(apertures)
-        apertures = [*zip(it,it)]
-        image = sparse_image(im, 900, 900)
-        
-        
-        fig, ax = plt.subplots()
-        image = color.gray2rgb(image)
-    
-        
-        for a in apertures:
-                ax.plot(a[0], a[1], color="darkred", marker='x', 
-                        linewidth=3, markersize=5)
-        
-        for b in balls:
-                ax.plot(b[0], b[1], color="darkblue",marker='o', 
-                        linewidth=3, markersize=2)
-        
-        ax.imshow(image)
-        ax.title.set_text(filename)
-        plt.show()
-        
-def plot_box(df):    
-    data_1 = (df['WL', 1,'x']**2*df['WL', 1,'y']**2)**(1/2)
-    data_2 = (df['WL', 2,'x']**2*df['WL', 2,'y']**2)**(1/2)
-    data_3 = (df['WL', 3,'x']**2*df['WL', 3,'y']**2)**(1/2)
-    data_4 = (df['WL', 4,'x']**2*df['WL', 4,'y']**2)**(1/2) 
-    
-    data = [data_1, data_2, data_3, data_4]
-    fig = plt.figure(figsize =(7, 5))
-     
-    # Creating axes instance
-    ax = fig.add_axes([0, 0, 1, 1])
-     
-    # Creating plot
-    bp = ax.boxplot(data)
-    ticks = plt.xticks([1,2,3,4],[30,50,60,100])
-    plt.title(label="MTSI Positional Verification", fontsize=25)
-    plt.xlabel("Distance from Isocentre (mm)", fontsize=10)
-    plt.ylabel("Deviation from DRR (mm)", fontsize=10)
-    
-    # show plot
-    plt.show()
-
-    
 def get_epid_balls_and_apertures(names, num_of_balls):
     """
-    Gets the ball and apeture positions from the EPID image.
+    Get the ball and apeture positions from the EPID image.
 
     Parameters
     ----------
@@ -287,7 +132,6 @@ def get_epid_balls_and_apertures(names, num_of_balls):
     None.
 
     """
-
     for i, n in enumerate(names):
         filename = epidfolder / n
         im = imageio.imread(filename)
@@ -323,14 +167,10 @@ def get_epid_balls_and_apertures(names, num_of_balls):
         
         # Progress bar
         update_progress(i/progmax, text)
-        
-        
-        
-
-
+      
 def get_drr_balls(names, num_of_balls):
     """
-    Gets the ball positions from the drr image.
+    Get the ball positions from the drr image.
 
     Parameters
     ----------
@@ -342,7 +182,6 @@ def get_drr_balls(names, num_of_balls):
     None.
 
     """
-
     for i, n in enumerate(names):
         filename = drrfolder / n
         im = imageio.imread(filename)
@@ -375,7 +214,7 @@ def get_drr_balls(names, num_of_balls):
 
 def get_drr_apertures(names, num_of_balls):
     """
-    Gets apeture positions from the mlc image.
+    Get apeture positions from the mlc image and saves to dataframe.
 
     Parameters
     ----------
@@ -389,8 +228,7 @@ def get_drr_apertures(names, num_of_balls):
     None.
 
     """
-
-    for i, n in enumerate(names):
+    for i, n in enumerate(names, df):
         filename = mlcfolder / n
         
         im = imageio.imread(filename)
@@ -441,8 +279,20 @@ def get_drr_apertures(names, num_of_balls):
         text = "Finding apertures in MLC {0}".format(n)
         update_progress(i/progmax, text)
 
-
 def calculateWL(df):
+    """
+    Calculate Winston Lutz deviations and saves to dataframe.
+
+    Parameters
+    ----------
+    df : pandas dataframe
+        dataframe of ball and aperture details.
+
+    Returns
+    -------
+    None.
+
+    """
     # calculate deviation in mm
     # Resolution details in DICOM header Image plane pixel spacing (3002,0011)
     # 0.336 mm/px at SID 100cm
@@ -462,11 +312,9 @@ def calculateWL(df):
     
     # remove values higher than 5 mm 
     df['WL'] = df['WL'][np.abs(df.WL) < 5]
-    
-    plot_against_gantry('WL')
-
 
 if __name__ == "__main__": 
+    
     fstring = 'P:/14 Projects/49_SRS Phantom/Output Images'
     data_folder = ('P:/14 Projects/49_SRS Phantom/Output Images')
     data_folder = Path(data_folder)
@@ -509,26 +357,30 @@ if __name__ == "__main__":
     get_drr_balls(names, num_of_balls)
     get_drr_apertures(names, num_of_balls)
     
-    plot_coords_on_images('epid', range(28,53)) 
+    calculateWL(df)
+    
+    plot_coords_on_images('epid', range(28,53), data_folder, df) 
+    plt.show()
+
     # Allowed arguments:'drrballs', 'drrape' 'epid' 
+       
+    
+    plot_against_gantry('DRRApertures', num_of_balls, df) # EPIDBalls, EPIDApertures, DRRBalls, DRRAperturs
+    plt.show()
+    plot_against_gantry('DRRBalls', num_of_balls, df)
+    plt.show()
+    plot_against_gantry('EPIDApertures', num_of_balls, df)
+    plt.show()
+    plot_against_gantry('EPIDBalls', num_of_balls, df)
+    plt.show()
         
     
-    plot_against_gantry('DRRApertures') # EPIDBalls, EPIDApertures, DRRBalls, DRRAperturs
-    plot_against_gantry('DRRBalls')
-    
-    plot_against_gantry('EPIDApertures')
-    plot_against_gantry('EPIDBalls')
-    
-    calculateWL(df)
-    # Scratch
+    plot_against_gantry('WL', num_of_balls, df)
+    plt.show()
+    boxplot('WL', num_of_balls, df)
+    plt.show()
 
-    plot_box(df)
-
-    #branch testing
-    #branch test 2
-
-
-
+    plt.close('all') # memory save
